add mailNewVote function
